Create FLEXi_Triggers
v 1.0

Use this SmartApp with FLEXiHue, FLEXiDimmer, and FLEXiState device types (and/or any kind of switch).  The FLEXi devices have custom attributes and commands that allow for the storage of scene data.  

To determine what color & level to use, this SmartApp applies the following logic:

1) If Scene Settings can be found, use those; 
2) If no Scene Settings, but Manual Settings can be found, use those; and
3) if neither Scene or Manual Settings can be found, then use the defaults entered in this App's preferences.
 
Scene settings for the FLEXi devices are selected and stored by the FLEXi Lighting SmartApp.  
Manual settings are made by selecting the individual device in the SmartThings App and changing any one of: level, hue, or saturation.